+++ src/tidop/img_process/color_convert.cpp
@@ -14,12 +14,12 @@ void rgbToCmyk(const cv::Mat &rgb, cv::Mat *cmyk)
   cmyk->create(rgb.size(), CV_32FC4);
   cv::Mat _cmyk = *cmyk;
 
-  parallel_for(0, rgb.rows, [&](int r) {
+  parallel_for(static_cast<size_t>(0), static_cast<size_t>(rgb.rows), [&](size_t r) {
     double cyan, magenta, yellow, key;
-    const uchar *rgb_ptr = rgb.ptr<uchar>(r);
+    const uchar *rgb_ptr = rgb.ptr<uchar>(static_cast<int>(r));
     for (int c = 0; c < rgb.cols; c++) {
       rgbToCmyk(rgb_ptr[3 * c + 2], rgb_ptr[3 * c + 1], rgb_ptr[3 * c], &cyan, &magenta, &yellow, &key);
-      _cmyk.at<cv::Vec4f>(r, c) = cv::Vec4f(static_cast<float>(cyan), static_cast<float>(magenta),
+      _cmyk.at<cv::Vec4f>(static_cast<int>(r), c) = cv::Vec4f(static_cast<float>(cyan), static_cast<float>(magenta),
         static_cast<float>(yellow), static_cast<float>(key));
     }
   });
@@ -32,14 +32,14 @@ void cmykToRgb(const cv::Mat &cmyk, cv::Mat *rgb)
   rgb->create(cmyk.size(), CV_8UC3);
   cv::Mat _rgb = *rgb;
 
-  parallel_for(0, cmyk.rows, [&](int r) {
+  parallel_for(static_cast<size_t>(0), static_cast<size_t>(cmyk.rows), [&](size_t r) {
     int red, green, blue;
     for (int c = 0; c < cmyk.cols; c++) {
-      cv::Vec4f v_cmyk = cmyk.at<cv::Vec4f>(r, c);
+      cv::Vec4f v_cmyk = cmyk.at<cv::Vec4f>(static_cast<int>(r), c);
       cmykToRgb(static_cast<double>(v_cmyk[0]), static_cast<double>(v_cmyk[1]), static_cast<double>(v_cmyk[2]), static_cast<double>(v_cmyk[3]), &red, &green, &blue);
-      _rgb.at<cv::Vec3b>(r, c)[0] = static_cast<uchar>(blue);
-      _rgb.at<cv::Vec3b>(r, c)[1] = static_cast<uchar>(green);
-      _rgb.at<cv::Vec3b>(r, c)[2] = static_cast<uchar>(red);
+      _rgb.at<cv::Vec3b>(static_cast<int>(r), c)[0] = static_cast<uchar>(blue);
+      _rgb.at<cv::Vec3b>(static_cast<int>(r), c)[1] = static_cast<uchar>(green);
+      _rgb.at<cv::Vec3b>(static_cast<int>(r), c)[2] = static_cast<uchar>(red);
     }
   });
 }
@@ -51,12 +51,12 @@ void rgbToHSL(const cv::Mat &rgb, cv::Mat *hsl)
   hsl->create(rgb.size(), CV_32FC3);
   cv::Mat _hsl = *hsl;
 
-  parallel_for(0, rgb.rows, [&](int r) {
+  parallel_for(static_cast<size_t>(0), static_cast<size_t>(rgb.rows), [&](size_t r) {
     double hue, saturation, lightness;
-    const uchar *rgb_ptr = rgb.ptr<uchar>(r);
+    const uchar *rgb_ptr = rgb.ptr<uchar>(static_cast<int>(r));
     for (int c = 0; c < rgb.cols; c++) {
       rgbToHSL(rgb_ptr[3 * c + 2], rgb_ptr[3 * c + 1], rgb_ptr[3 * c], &hue, &saturation, &lightness);
-      _hsl.at<cv::Vec3f>(r, c) = cv::Vec3f(static_cast<float>(hue), static_cast<float>(saturation), static_cast<float>(lightness));
+      _hsl.at<cv::Vec3f>(static_cast<int>(r), c) = cv::Vec3f(static_cast<float>(hue), static_cast<float>(saturation), static_cast<float>(lightness));
     }
   });
 }
@@ -68,14 +68,14 @@ void hslToRgb(const cv::Mat &hsl, cv::Mat *rgb)
   rgb->create(hsl.size(), CV_8UC3);
   cv::Mat _rgb = *rgb;
 
-  parallel_for(0, hsl.rows, [&](int r) {
+  parallel_for(static_cast<size_t>(0), static_cast<size_t>(hsl.rows), [&](size_t r) {
     int red, green, blue;
     for (int c = 0; c < hsl.cols; c++) {
-      cv::Vec3f v_hsl = hsl.at<cv::Vec3f>(r, c);
+      cv::Vec3f v_hsl = hsl.at<cv::Vec3f>(static_cast<int>(r), c);
       hslToRgb(static_cast<double>(v_hsl[0]), static_cast<double>(v_hsl[1]), static_cast<double>(v_hsl[2]), &red, &green, &blue);
-      _rgb.at<cv::Vec3b>(r, c)[0] = static_cast<uchar>(blue);
-      _rgb.at<cv::Vec3b>(r, c)[1] = static_cast<uchar>(green);
-      _rgb.at<cv::Vec3b>(r, c)[2] = static_cast<uchar>(red);
+      _rgb.at<cv::Vec3b>(static_cast<int>(r), c)[0] = static_cast<uchar>(blue);
+      _rgb.at<cv::Vec3b>(static_cast<int>(r), c)[1] = static_cast<uchar>(green);
+      _rgb.at<cv::Vec3b>(static_cast<int>(r), c)[2] = static_cast<uchar>(red);
     }
   });
 }
@@ -131,12 +131,12 @@ void chromaticityCoordinates(const cv::Mat &rgb, cv::Mat *chroma_rgb)
   chroma_rgb->create( rgb.size(), CV_32FC3);
   cv::Mat chroma_coord = *chroma_rgb;
 
-  parallel_for(0, rgb.rows, [&](int r) {
+  parallel_for(static_cast<size_t>(0), static_cast<size_t>(rgb.rows), [&](size_t r) {
     double c_red, c_green, c_blue;
-    const uchar *rgb_ptr = rgb.ptr<uchar>(r);
+    const uchar *rgb_ptr = rgb.ptr<uchar>(static_cast<int>(r));
     for (int c = 0; c < rgb.cols; c++) {
       chromaticityCoordinates(rgb_ptr[3*c+2], rgb_ptr[3*c+1], rgb_ptr[3*c], &c_red, &c_green, &c_blue);
-      chroma_coord.at<cv::Vec3f>(r, c) = cv::Vec3f(static_cast<float>(c_blue), static_cast<float>(c_green), static_cast<float>(c_red));
+      chroma_coord.at<cv::Vec3f>(static_cast<int>(r), c) = cv::Vec3f(static_cast<float>(c_blue), static_cast<float>(c_green), static_cast<float>(c_red));
     }
   });
 }
